Extract async sibling accessibility analyzer
